Fix camera snapping to correct position at end of animation
This was due to the planet orbit causing the end point of the animation being different than the planet's original location when starting the animation. By using the pointer to the planet position, we are able to animate to the current position, not the original planet position at the time of mesh picking.